Switch is_directed_acyclic_graph to use toposort internally
Previously the retworkx.is_directed_acyclic_graph() function was
internally using the petgraph::algo::is_cyclic_directed function to look
for a potential cycle in the graph. However, as noted in the docs[1]
this function is recursive and as reported in Qiskit/qiskit-terra#5502
with very large graphs this is causing segfaults. To avoid this
potential issue this commit changes the internal implementation to use
petgraph::algo::toposort which isn't recursive and should be more
reliable.

Fixes: Qiskit/qiskit-terra#5502

[1] https://docs.rs/petgraph/0.5.1/petgraph/algo/fn.is_cyclic_directed.html