Update documentation and add release notes (#214)
* Update documentation and add release notes

This commit updates the documentation theme to match the current theme
that the qiskit project is using. It also takes the opportunity to add
the release notes to the documentation. We previously were only
publishing the release documentation on the github release page which
has limitations. Moving forward we'll use reno to automate the
generation of release notes.

* Use a full clone in docs job for reno

* Finish converting existing release documentation to sphinx rst

* Update contributing guide with reno details

* Add contributing guide to the documentation

* Add note about rustfmt to tox lint contributing section

* Apply suggestions from code review

Co-authored-by: Corey Mendell <corey.mendell@gmail.com>

* Rework release note section titles to be consistent with reno's section titles

* Add contributing to sdist to fix CI job

* Add missing contributing rst file

Co-authored-by: Corey Mendell <corey.mendell@gmail.com>