Fix the read the documentation builds (#218)
In #214 we updated the documentation configuration to use the new
qiskit sphinx theme to be consistent with qiskit's documentation and
also add the use of reno for release notes. However, that PR did not
update the conda environment file used by read the docs for buidling the
documetation to include the new requiremetns. This commit corrects that
oversight and adds the missing dependencies to the docs builds.