Fix most of the test panics
This commit fixes all but 2 of the test panics. There were 2 classes of
issues that his commit fixes (both were artifacts of sloppy porting from
Python). The first is that in add_blossom() we were not updating the
global blossom_children or the global blossom_endpoints while iterating
over the list. Instead we were just setting a local. The second issue
was negative indices for Vecs. The python code was iterating over
elements in a vec in expand_blossom() and augment_blossom() by
subtracting a length or step from the index which could result in a
negative value. In Python this is fine since a negative index is just
the from the last element. But, for rust we need to handle this behavior
manually because the Index trait only deals with usize.