Avoid allocation overhead with_capacity() instead of new()
This commit replaces uses of ::new() for container types (like Vec and
HashMap) with with_capacity(). The difference between the 2 is that
with_capacity() makes the initial memory allocation with N elements of
the given type in the container (and still grows when it's exceeded)
while new() just makes a minimal initial memory allocation and relies on
more allocations. So it becomes a tradeoff between using potentially
excess memory vs th overhead of additional memory allocations. So for
cases where we have an idea of the final size of the container object it
makes sense to avoid the memory allocation and just do a large one up
front.