Fix handling of holes in node indices and add test
This commit fixes an issue when running max_weight_matching on a graph
with holes in the node indices (i.e. when
'node_count() != max(node_indices)'). The in_blossoms vec was being
initialized from the contents of the node_indices iterator of the graph
but the algorithm remaps the indices to be a contiguous list for the
internals to work. When a hole was encountered accessing in_blossoms
would panic because the index returned wouldn't match the expecated
state. This is fixed and a test added to make sure we have coverage on
input graphs with holes.