Avoid allocation overhead with_capacity() instead of new() (#233)
This commit replaces uses of ::new() for container types (like Vec and
HashMap) with with_capacity(). The difference between the 2 is that
with_capacity() makes the initial memory allocation with N elements of
the given type in the container (and still grows when it's exceeded)
while new() doesn't allocate until something is added and relies on
more allocations as data is added. [1] So it becomes a tradeoff between
potentially allocating excess memory vs the overhead of additional
memory allocations. So for cases where we have an idea of the final
size of the container object it makes sense to avoid the memory
reallocation and just do a large one up front.

[1] https://doc.rust-lang.org/std/vec/struct.Vec.html#capacity-and-reallocation