Use full LTO for release builds (#232)
This commit changes the release build configuration to use full LTO for
retworkx release builds. The tradeoff this makes is compile time vs
optimization. It basically serializes the compilation of retworkx's
dependencies so that it can do link time optimization. From benchmarking
the performance gains are modest (in the best case ~2-5%), but it
still seems worth it since most users don't constantly rebuild retworkx
and just use the prebuilt binaries that are included in the wheels on
PyPI so tuning the release builds at the cost of compile time seems
worth it.