Add APIs to update an edge's data payload inplace
This commit adds 2 new methods to the graph classes, update_edge and
update_edge_by_index, which are used to update the data payload/weight
of an edge in the graph in place. Prior to this the only way to update
an edge would be to remove it and add a new edge with the new data
payload which may change the edge index (assuming something was using
it).