Revert "Bump pyo3 from 0.12.4 to 0.13.1 (#224)" (#237)
This commit reverts the version bumps for pyo3, numpy, and ndarray. The
3 changes here required increasing the minimum supported rust version
from 1.39.0 to 1.45.0. While for most developers and use cases this is
fine since 1.45.0 is > 6 months old and most packagers/distributors keep
the rust version up-to-date. One known exception to this though is the
debian 10 rustc package is at 1.41.0. This becomes an issue for
raspberry pi users running a 32bit operating system (the aarch64
manylinux wheels on PyPI work fine for 64bit OS rasberry pi users).
The piwheels project is using the debian 10 package which means to
provide prebuilt wheels of retworkx they need to be able to use rustc
1.41.0 to build.

When weighing the features of this version bump (the big thing for our
use case are better type error messages when calling retworkx from
python) that this version bump would require all raspberry pi users
to switch back to building retworkx from source on 32bit Raspberry Pi OS
when retworkx on piwheels was only a recent addition we should wait at
least one more release before bumping the minimum support rust version.
Hopefully the debian rustc version will be bumped to >=1.45.0 at some
point soon and then this issue goes away.

This reverts commit a3249f0825028c1799984547e9c62bcd2cef89ab.