Fix rust lint issues (#244)
This commit fixes some small rust lint issues caught with the nightly
version of clippy (rust 1.51.nightly) which has some additional rules
enabled. The vec![] usage should have a minor performance improvement
over the Vec::new() and push() that was used before but it is likely not
noticeable in practice.