Fix s390x release wheel job (#245)
This commit fixes the s390x wheel jobs. In the 0.7.2 release the s390x
wheel job failed while timing out waiting for data. Running the job
locally it is hanging on the yum clean up steps in the script to install
rust on s390x. This was needed at one time in the past because the
available disk space on the s390x travis CI nodes used to be exceedingly
small and just downloading the rust compiler would use up all the free
space, so we went to all effort to free up space so we could build the
wheels. However, those issues in the s390x CI nodes have been fixed and
we're no longer under a tight space constraint. This commit just removes
these steps which were hanging and bumps the rust version used to the
latest, 1.49.0. This will fix the jobs for when we push the next release
(a 0.7.2 build was manually retriggered with this applied and has
already been published to PyPI).