Add release job to publish macOS arm64 binaries (#246)
* Add release job to publish macOS arm64 binaries

With the recent release of cibuildwheel 1.9.0 support for building arm64
binaries on macOS was added. This enables us to build release binaries
for this new platform. However, these binaries will be cross-compiled
and not tested since there are no CI resources with the platform
available. This commit adds a new release wheel job to build a aarch64
binary and publish an arm64 wheel for macOS. This will enable users who
have an arm64 mac to run retworkx without having to either build from
source or run under rosetta 2.

Fixes #231

* DNM: Testing the wheel builds

* Install aarch64-apple-darwin target

* Random guess at cross lib dir

* Update PYO3_CROSS_LIB path

* Revert "DNM: Testing the wheel builds"

Tested the wheel builds work and @nonhermitian confirmed the built
wheels worked on an arm64 mac locally.

This reverts commit 2a6ff1fd8d2d34fe4217c09042dda6ebc099c5a0.

* Install twine in arm64 wheel job

* Add release notes