Add back verify_optimum() function and make it an optional flag
This commit adds back the verify_optimum solution to check that the
found matching is a optimum solution. An earlier draft of this function
was included in prior commits but was removed because it was never run.
This fixes the issues with the earlier draft and also makes it an opt-in
option for users. By default it is not run, but if a user specifies they
want the verification run there is a flag on the python API function to
enable it. This flag is then used on all the tests calls (except where
it's explicitly tested as False to exercise that code path).