Fallback to checking valid matching and optimum result
In the unit tests that compare the output of running against networkx if
there are multiple valid maximum weight matchings in the random graph
the results between networkx and retworkx can differ but both still be
valid. This commit modifies the tests to validate in those cases that
the output from retworkx is still a valid maximal matching and relying
on the verify_optimum flag to ensure the output result is optimal we
should have confidence that the result is valid.