Ensure uniqueness of neighbor indices returned by `neightbors()`
petgraph's underlaying `neighbors()` method returns duplicated
nodes in multigraphs (on per-edge basis). The desired behavior
for the public API is not to include these repetitions though.

This commit collects the returned indices in a HashSet to remove
duplications, then it exhausts the set with `drain()` and
collect it into a list.

This commit also adds new tests to check there is no duplication
in the list of neighbors indices and adapt the existing tests to
ignore the order.