Pin rust in s390x test job (#254)
* Pin rust in s390x test job

Since the Rust 1.50.0 the s390x CI test job has been failing. This looks
to be caused by rust-lang/rust#80810. Until the issue is resolved in a
released version of rust this pins the rust version used in the job to
the previous release 1.49.0 which does not have this issue and should
work fine.

* Call rustup in the before_install stage