Suppress new clippy warning in rust 1.50.0
In rust 1.50.0 the new version of cargo clippy is highlighting a warning
on an unecessary result on the output for a closure, since the closure
is always returning true (so there is no possibility of an error output.
However, this is ignoring that the closure is passed to a function and
the return type expected by that function has a Result<> return. If we
were to remove the result from the closure that would become a type
mismatch. To avoid clippy raising an issue in this instance this commit
adds allow on that lint rule for that line.