Fix issues identified by newer cargo clippy ersions
This commit fixes several instances when we were using the '?' to
return an error result in a return line. For example something like:
return 'Ok(fn()?)'. This was unecessary because the output type is
Result if we're using the '?' so we can just pass the result directly
instead of wrapping the output in Ok().