Fix issues found with newer cargo clippy versions (#257)
* Suppress new clippy warning in rust 1.50.0

In rust 1.50.0 the new version of cargo clippy is highlighting a warning
on an unecessary result on the output for a closure, since the closure
is always returning true (so there is no possibility of an error output.
However, this is ignoring that the closure is passed to a function and
the return type expected by that function has a Result<> return. If we
were to remove the result from the closure that would become a type
mismatch. To avoid clippy raising an issue in this instance this commit
adds allow on that lint rule for that line.

* Fix issues identified by newer cargo clippy ersions

This commit fixes several instances when we were using the '?' to
return an error result in a return line. For example something like:
return 'Ok(fn()?)'. This was unecessary because the output type is
Result if we're using the '?' so we can just pass the result directly
instead of wrapping the output in Ok().

* Fix new clippy issues in max_weight_matching