Revert "Try using arm64 graviton2 instead of arm64 (#253)"
The arm64-graviton2 nodes are not being correctly allocated and are
showing as an x86_64 architecture (it looks like they're not even
running on amazon ec2). Since this defeats the purpose of the arm64
ci jobs (to test on arm64) this commit reverts the change to graviton2
and uses the regular slower arm64 nodes, which are actually run on arm
cpus.

This reverts commit bbcb0d55c1dbb0c575b1be3a29a9594e3bb68a14.