Add more object protocol methods to custom return types
This commit adds a few more methods from the Python object protocol to the
custom return types. All the classes now have a __str__ method so you can
see the contents of the custom returns with a print(). Additionally, a
__hash__ is added to each class since the return types are immutable.
However, for return types that contain PyObjects (WeightedEdgeList and
BFSSuccessors) if any of the objects contained are not hashable this
will raise an exception (just as if you called hash() on that object).
The last methods are related to garbage collection in python which
enables the Python garbage collector to cleanup the custom return types.