Simplify is_matching logic
The original implementation was based on networkx's which was getting
all the combinations of edges in the matching and checking for unique
endpoints on all those pairs. However, doing the combinations adds
extra complexity when we can just flatten the input matching to a set of
endpoints and if aduplicate is found it's not a valid matching. This
commit updates the inner is_matching function to make this change which
simplifies the code and should be faster.

Co-authored-by: Toshinari Itoko <itoko@jp.ibm.com>