Add is_maximal_matching and is_matching functions (#256)
* Add is_maximal_matching and is_matching functions

This commit adds new functions for checking a provided matching set is
valid, is_matching(), and that a provided matching set is valid and
maximal, is_maximal_set(). This pairs with #229 and can be used to
partially check the output from the max_weight_matching function added
there. Equivalent functions were implemented in #229 using Python for
the tests in #229 and those tests should be updated to use these
functions instead.

Fixes #255

* Use retworkx functions in max_weight_matching tests

* Remove unused import

* Simplify is_matching logic

The original implementation was based on networkx's which was getting
all the combinations of edges in the matching and checking for unique
endpoints on all those pairs. However, doing the combinations adds
extra complexity when we can just flatten the input matching to a set of
endpoints and if aduplicate is found it's not a valid matching. This
commit updates the inner is_matching function to make this change which
simplifies the code and should be faster.

Co-authored-by: Toshinari Itoko <itoko@jp.ibm.com>

* Update docstring

Co-authored-by: Toshinari Itoko <itoko@jp.ibm.com>