Fix CI wheel jobs (#262)
* Fix CI wheel jobs

For the recent 0.8.0 release all the wheel jobs failed because we added
networkx as a test requirement (for comparing the outputs of
max_weight_matching()) but were not installing networkx before running
unit tests. This made all the wheels come back as broken because
unittests failed. This commit fixes that issue by adding networkx as an
explicit test requirement in all the cibuildwheel job configurations.

Additionally, the ppc64le wheel jobs started failed for an additional
reason, it appears that the travis CI images aren't up to date and need
the rust compiler to install cryptography (which is required for pip and
other packages). This commit fixes those jobs so we can install
cibuildwheel and also build the packages.

* Fix ppc64le job

* Make sdist job depend on wheel jobs

* DNM: test wheel jobs

* Revert "DNM: test wheel jobs"

This reverts commit 753a739f55b42176a50fe5f359d904c20008e843.