Use stestr for test jobs
Since #229 our test runs take significantly longer. This is mostly a
function of the tests added in that function running over numerous
random graphs and comparing the output against networkx to verify it
works in all cases. However, this validation comes with a runtime cost.
In an attempt to mitigate this partially this commit changes the test
runner from stdlib's unittest runner to use stestr. [1][2][3] stestr
provides a performance improvement by running tests on parallel workers
and also provides an easier to use UI that enables test selection a
history of previous runs, and many other features. The test suite in
retworkx will still need to be strictly unittest compatible so anyone
can run with stestr, stdlib unittest, pytest, or any other test runner
of their choice. But the default in tox and CI will be to use stestr
for the performance benefit.

[1] https://pypi.org/project/stestr/
[2] https://github.com/mtreinish/stestr
[3] https://stestr.readthedocs.io/en/stable/README.html