Use std::option::Option::match where applicable
There was a common pattern in places of the retworkx code that would do
something like:

match foo {
    Some(foo) => bar(foo),
    None => None,
}

which is essentially the same thing that Option.map() does. [1] To
reduce the code verbosity this commit goes through and updates the
places that we were doing Option.map() manually like this and replaces
it with a map() call.

[1] https://doc.rust-lang.org/std/option/enum.Option.html#method.map