Use rust-numpy 0.13.1 and bumpy MSRV to 1.41.1
As a conclusion to PyO3/rust-numpy#146 the MSRV for rust-numpy was
defined as matching PyO3 and being 1.41.1. The rust-numpy 0.13.1
release handled this and relaxed it's dependency on upstream
dependencies (namely ndarray) to enable building with rust 1.41.1. Since
rust 1.41.1 is what is shipped with Debian buster it will work for
retworkx (so we don't block piwheels which relies on the debian rustc
package). This commit bumps the rust-numpy version to the latest release
and correspondingly bumps the MSRV to match at 1.41.1 which also unblocks
the use of the latest PyO3 release.