Revert "Revert "Bump pyo3 from 0.12.4 to 0.13.1 (#224)"" (#238)
* Revert "Revert "Bump pyo3 from 0.12.4 to 0.13.1 (#224)""

This commit restores the PyO3, rust-numpy, and ndarray bump to the
latest versions. This was previously reverted in #237 to avoid the
increase in the minimum supported rust version to enable builds on
piwheels. However, now that 0.8.0 has been released I think it's fine to
bump now.

This reverts commit 3beb8fbc8b3794a13009bedca1a43c4b524b9570.

* Relax msrv to 1.42.0

* Update bump-msrv-d581a8cbeff8915b.yaml

* Update README.md

* Use rust-numpy 0.13.1 and bumpy MSRV to 1.41.1

As a conclusion to PyO3/rust-numpy#146 the MSRV for rust-numpy was
defined as matching PyO3 and being 1.41.1. The rust-numpy 0.13.1
release handled this and relaxed it's dependency on upstream
dependencies (namely ndarray) to enable building with rust 1.41.1. Since
rust 1.41.1 is what is shipped with Debian buster it will work for
retworkx (so we don't block piwheels which relies on the debian rustc
package). This commit bumps the rust-numpy version to the latest release
and correspondingly bumps the MSRV to match at 1.41.1 which also unblocks
the use of the latest PyO3 release.