Add networkx converter function
A commonly asked for piece of functionality of existing networkx users
is an easy way to convert networkx objects into retworkx objects. While
in the past we've avoided this because there is a strong desire to
**never** add a dependency on networkx to retworkx implementing a
converter from networkx could be built by just duck typing and never
needing to import networkx. This commit implements that function so that
if a user of retworkx is taking in networkx graphs they have a simple
function to get a retworkx graph object from that. We still will never
add networkx as a dependency, which means going from retworkx graphs to
networkx graphs will be an exercise for users. But this commit also adds
a code example on how to do this to the retworkx for networkx users
guide to hopefully make that simpler.