Add the rest of dict methods to custom return classes
Since an extension class implementing the Python Mapping protocol
doesn't get full dictionary compatibility just by implementing the
protocol we need to add additional methods and iterator support
manually. This commit does this for the new PathMapping and
PathLengthMapping return types. It adds new view classes for keys,
values, and items and then adds the corresponding keys(), values(),
items(), and __iter__ support to use those. Then to ensure that
__contains__ is implemented (eg so you can do x in PathMapping) the
sequence protocol needs to be implemented because that is where the C
api in python defines the necessary methods.