Pin decorator for CI and testing
With the recent release of decorator 5.0.x CI has been failing. This is
because networkx, which is a test requirement, was broken by changes in
that release. Since we don't directly depend on decorator this commit
adds a constraints file and pins the decorator version to the last known
working version.