Add matplotlib based drawer
This commit adds a a new python visualization module which contains a
matplotlib drawer for retworkx graphs. This first iteration of the
visualization function is porting networkx's equivalent functionality.
For now the api is basically identical to networkx's functionality
except it takes a retworkx graph object instead of a networkx graph
object. However, in follow-on commits to this branch the API will be
adapted to be more retworkx native (i.e. using callbacks to generate
labels from nodes and edges, etc).

To implement this a basic random_layout function is added to retworkx
and a corresponding 2D layout type Pos2DMapping which is used to
return the layouts efficiently (instead of creating a dict).

This is still a rough draft and before this merges into master things
will be updated.

Fixes #298