Add random_layout function and Pos2DMapping
This commit adds a new function random_layout() which is used to
generate a random layout for a graph that can be used in visualization.
This is necessary for building a matplotlib drawer (issue #298 and a
first draft of the implementation #304). To make the function more
efficient it also adds a new custom return type Pos2DMapping which is
used to build an imutable readonly dict compatible result container for
the output type from this function.

Related to #280