Reorganize into visualization module and cleanup API
This commit starts the process of reorganizing the matplotlib drawer
into the final form. It flattens the api so there is a single public api
entrypoint, mpl_draw() which exposes all the drawer options and returns
a figure (if not in interactive mode). This is the first module and
function in the visualization subpackage. In the future we will add some
other ones built around graph.to_dot() that use pydot (so we avoid the
boilerplate drawer code everywhere).