Skip rustup installation for ppc64le CI job
The ppc64le CI job has started to fail recently when installing rustup
to bootstrap a rust environmnet in the CI node. The error states that
there is already a system rust installation and that it will conflict
with rustup's version of rust. Since apparently rust is preinstalled on
the travis ppc64le CI nodes now this commit attempts to fix the error
by skipping the rustup step and using the preinstalled rust compiler.