Use [f64; 2] for center type
This commit changes the input type of the center option to be a fixed 2
element array instead of a tuple. This still works as a tuple for the
python interface but has a defined contiguous memory layout in rust. It
also is consistent with the type for elements in the position mapping.