Add random_layout function and Pos2DMapping (#305)
* Add random_layout function and Pos2DMapping

This commit adds a new function random_layout() which is used to
generate a random layout for a graph that can be used in visualization.
This is necessary for building a matplotlib drawer (issue #298 and a
first draft of the implementation #304). To make the function more
efficient it also adds a new custom return type Pos2DMapping which is
used to build an imutable readonly dict compatible result container for
the output type from this function.

Related to #280

* Add release note

* Update doc organization

* Apply suggestions from code review

Co-authored-by: Jielun (Chris) Chen <31495624+Chriscrosser3310@users.noreply.github.com>

* Update releasenotes/notes/add-random-layout-c1c2751be971e5d0.yaml

Co-authored-by: Nahum Rosa Cruz Sa <49600259+nahumsa@users.noreply.github.com>

* Update releasenotes/notes/add-random-layout-c1c2751be971e5d0.yaml

Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>

* Use [f64; 2] for center type

This commit changes the input type of the center option to be a fixed 2
element array instead of a tuple. This still works as a tuple for the
python interface but has a defined contiguous memory layout in rust. It
also is consistent with the type for elements in the position mapping.

Co-authored-by: Jielun (Chris) Chen <31495624+Chriscrosser3310@users.noreply.github.com>
Co-authored-by: Nahum Rosa Cruz Sa <49600259+nahumsa@users.noreply.github.com>
Co-authored-by: georgios-ts <45130028+georgios-ts@users.noreply.github.com>