Fix mutability of self on successor/predecessor indices methods
This commit fixes a small overight in the predecessor_indices() and
successor_indices() methods for PyDiGraph. These functions were
previously set so they get a mutable borrow of self, but did not mutate
the contents of the graph object. This caused potential issues if/when
these methods were used in a context where there was another reference
to the graph object (like if they were used in the callback nethod for
a graph method) that would raise an error when the method was called.
This commit corrects this oversight by correctly setting the reference
to self to be immutable.