Add custom returns types for path dicts and path length dicts (#287)
* Add custom returns types for path dicts and path length dicts

This commit adds 2 new custom return types PathMapping and
PathLengthMapping. These are used for functions that return either a
path dictionary or path length dictionary. These return types implement
the python mapping protocol and should be usable inplace of the previous
dict replacement.

* Fix build on MSRV and add missing releasenote

* Add test to cover index errors in new types

* Add the rest of dict methods to custom return classes

Since an extension class implementing the Python Mapping protocol
doesn't get full dictionary compatibility just by implementing the
protocol we need to add additional methods and iterator support
manually. This commit does this for the new PathMapping and
PathLengthMapping return types. It adds new view classes for keys,
values, and items and then adds the corresponding keys(), values(),
items(), and __iter__ support to use those. Then to ensure that
__contains__ is implemented (eg so you can do x in PathMapping) the
sequence protocol needs to be implemented because that is where the C
api in python defines the necessary methods.

* Fix lint

* Fix test organization

* Fix lint

* Update releasenotes/notes/custom-path-returns-e1d5fc1e02809179.yaml

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Run rustfmt after rebase

* Fix lint