Allow failures on ppc64le linux travis jobs
In recent weeks we've seen a spike in job timeouts and other spurious
failures for the ppc64le CI. It looks like the travis ppc64le
environment is facing some issues. Since none of the recent failures
have caught any real code issues when running or building on ppc64le
and are all caused by infrastructure issues, this commit adds it to
the allowed failure list. We can remove it when and if the ppc64le ci
environment stabilizes.