Add write_edge_list methods (#309)
* Add write_edge_list methods

In #113 we added a constructor method to the graph classes to create a
graph object from a edge list file. However there was not a native way
to generate such a file from retworkx. This commit adds that function
with a new method, write_edge_list(), which can be used to write an edge
list file in the same format that can be read by read edge list.

* Fix lint

* Add explicit writer flush before dropping

Per the documentation of BufWriter struct [1] it is necessary to
manually flush the buffer prior to dropping the struct. This commit adds
that so the contents are always committed to the file prior to the
functions exiting.

[1] https://doc.rust-lang.org/std/io/struct.BufWriter.html