Add matplotlib based drawer (#304)
* Add matplotlib based drawer

This commit adds a a new python visualization module which contains a
matplotlib drawer for retworkx graphs. This first iteration of the
visualization function is porting networkx's equivalent functionality.
For now the api is basically identical to networkx's functionality
except it takes a retworkx graph object instead of a networkx graph
object. However, in follow-on commits to this branch the API will be
adapted to be more retworkx native (i.e. using callbacks to generate
labels from nodes and edges, etc).

To implement this a basic random_layout function is added to retworkx
and a corresponding 2D layout type Pos2DMapping which is used to
return the layouts efficiently (instead of creating a dict).

This is still a rough draft and before this merges into master things
will be updated.

Fixes #298

* Reorganize into visualization module and cleanup API

This commit starts the process of reorganizing the matplotlib drawer
into the final form. It flattens the api so there is a single public api
entrypoint, mpl_draw() which exposes all the drawer options and returns
a figure (if not in interactive mode). This is the first module and
function in the visualization subpackage. In the future we will add some
other ones built around graph.to_dot() that use pydot (so we avoid the
boilerplate drawer code everywhere).

* Remove pointless retworkx_ prefix and add edge label options

* Add tests

* Relax matplotlib requirement

* Fix mpl<3.4 compat

* Fix test skip condition

* Add release notes

* Run black

* Update default layout method to spring layout