Add pydot drawer function
This commit expands the retworkx.visualization module to include a new
pydot based drawer. This function wraps the pydot and pillow libraries
to generate graph visualizations leveraging graphviz. It internally uses
the to_dot() method of the retworkx graph types and then passes that dot
file to pydot to generate a visualization with graphviz. The result is
wrapped as a pillow image object.

Fixes #321