Bump dependencies in Cargo.lock file (#326)
Since we're planning to release later this month, this commit bumps the
dependency versions in the Cargo.lock to the latest versions for
everything where we're not constrained by the MSRV, mainly ndarray and
num-complex (which cargo tries to update as deps of numpy). This will
likely be needed again right before the release, but at least in the
meantime we can get some more testing in with the latest versions of
the dependencies before that.