Add matplotlib to read the docs builds
Since #304 merged building the retworkx documentation has required that
matplotlib be installed for generating the matplotlib visualizations in
the documentation of `mpl_draw()`. However that PR neglected to add it
to the read the docs conda environment configuration. This commit
corrects this oversight to hopefully fix the documentation builds on
read the docs.