Fix read the docs builds and docs formatting (#329)
* Add matplotlib to read the docs builds

Since #304 merged building the retworkx documentation has required that
matplotlib be installed for generating the matplotlib visualizations in
the documentation of `mpl_draw()`. However that PR neglected to add it
to the read the docs conda environment configuration. This commit
corrects this oversight to hopefully fix the documentation builds on
read the docs.

* Fix docs theme formatting

This commit fixes some docs theme formatting issues to make sure the
compiled HTML docs are correctly formatted. It mainly adds missing
template files and configuration so the qiskit-sphinx-theme package has
all the necessary details to correctly render.

* Update copyright string