Add `find_successors_by_edge` (#319)
Related to #265

This PR adds two new functions to PyDiGraph: find_sucessors_by_edge and
find_predecessors_by_edge. These functions can be useful for speeding up routines that
want a subset of successors/predecessors that match a filter function.

With these new functions, it is possible to find the successors/predecessors of an node that
match a filter in O(e'), where e' is the number of edges connected to the node. Before, the
simplest way to achieve the same behavior was to get all edges for each successor/predecessor
and then filter, which was O(d * e'), where d is the number of distinct neighbors a node has.

These new functions will be immediately applicable to the quantum_successors and
quantum_predecessors methods in qiskit-terra's DAGCircuit class. They will consequently provide a
speed up to collect 2q blocks, which has quantum_successors and quantum_predecessors as its
most expensive steps during graph traversal.

* Add filtered_successors/filtered_predecessors

* Add release note

* Rename new functions for consistency

* Add test cases for new functions

* Release note updates from code review

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>