Add missing doc theme template file (#331)
This commit adds a missing template file for the qiskit-sphinx-theme.
PR #329 previously attempted to fix the RTD docs builds and update the
theme formatting. While that worked fine for PR mode after it merged the
builds for the latest version of the docs did not work and fail with an
error about the missing template file. This will hopefully fix the
builds so we can see the rendered docs for the current state of the main
branch.