Migrate docs hosting from read the docs to qiskit.org/documentation
This commit migrates the documentation hosting from read the docs to
qiskit.org/documentation/retworkx. This will give us more flexibility in
how we build the documentation because it's just a static host in IBM's
cloud object storage so any static content we upload can be hosted and
we don't have to worry about the constraints on read the docs build
environment anymore, which was proving to be a bit limiting.