Add stable versions to the documentation (#334)
* Add stable versions to the documentation

This commit adds support for keeping multiple released versions of the
documentation on the documentation page. It will keep a single version
of the doc for each minor release of retworkx (for example 0.8, 0.9,
etc). Each tag documentation build will upload it's built docs to both
the latest stable url and the appropriate url on stable/. The version
table for the dropdown will also be populated with the previous versions
from 0.8 to current.

* Update tools/deploy_documentation.sh