Fix typo in version table
In #334 we added support for building and hosting stable versions of the
documentation at each minor version release. However, in that PR the
links added via the jinja template to versions.html had a small typo so
they didn't actually end up as proper html hyperlinks. This commit
corrects the oversight so that when the documentation is built the
version dropdown correctly has a link to the older versions of the
documentation.