Correct target for version drop down links
In #334 we added support for multiple versions of the documentation.
However that PR has a small issue in the target location of the links
for previous release section of the versions drop down. The urls it
generated result in a 404 because they're not actually targeting the
proper endpoint for the hosted documentation. This commit fixes that so
when the links are clicked it goes to the old version of the
documentation.