Add frozen attribute to graph classes
This commit adds a new attribute to a graph class, frozen, which when
set to True will make a graph object immutable. When it's set to True
all methods that mutate a graph object raise an exception. The one
exception is that all weight/data payload objects are stored and
returned by reference in the graph objects and if they're modified
in python there is no way to control that from retworkx.