Add methods to get number of nodes and edges in the graph (#335)
This commit adds two new methods num_nodes() and num_edges() to the
graph classes for returning the number of nodes and edges respectively
in the graph. While a user could do len(graph) to get the number of
nodes having an explicit method also is useful as it might not be
obvious that you can use len() on the graph.